feat(cred): resolve claim templates in an aws secret_id
A stored-secret id can now carry {{user.<claim>}} and {{agent.<claim>}} tokens,
so one spec serves many callers and each reads only its own secret, the way a
KV path already could.

The mechanism moves out of the vault driver into its own file first. It was
never vault-specific — the transit signer already used it for two more fields —
and a second driver reaching into a neighbour for it is the point at which that
stops being tidy. Pure move: same package, same unexported names, no caller
changed. Its tests move with it, since what they cover is the rule set rather
than the one caller they happen to go through. The doc comments lose their
references to one particular client's path handling; the rule they describe was
always general.

Resolution happens inside the shared fetch, so both mint methods that read a
stored secret get it, and the one being added next inherits it without knowing.
The non-exchange path passes no claims, which makes a templated id fail closed
there rather than be sent as written — that matters more than it looks: a store
will return a secret genuinely named "prod/{{user.sub}}/keys" to whoever can
create one.

The assertion's resource claim keeps carrying the unresolved id, as every
templated coordinate in that layer does — it is built before the exchange that
would produce the claims. So a downstream policy conditioning on it pins the
spec rather than the secret, and per-principal scoping is left to the
permissions on the assumed role, which is where the resolved read happens.
